New: Add support for older devices.
Some devices may not have the java crypto libraries available, particularly on Android API versions lower than 23. Unfortunately, this can render a wallet unable to be used on those devices because keys cannot be generated. This PR changes that and allows fall back crypto implementations to be used whenever the defaults are missing.